split frontpage by function
