--- .nyc_output/0961476f-548a-483e-bdcb-44b14ef1203d.json
@@ -1,1 +1,0 @@
-{"/usr/local/splash/models/metaData.js":{"path":"/usr/local/splash/models/metaData.js","statementMap":{"0":{"start":{"line":1,"column":20},"end":{"line":1,"column":388}},"1":{"start":{"line":3,"column":18},"end":{"line":30,"column":1}},"2":{"start":{"line":4,"column":22},"end":{"line":4,"column":62}},"3":{"start":{"line":5,"column":11},"end":{"line":5,"column":31}},"4":{"start":{"line":6,"column":12},"end":{"line":6,"column":27}},"5":{"start":{"line":7,"column":1},"end":{"line":7,"column":150}},"6":{"start":{"line":8,"column":1},"end":{"line":8,"column":12}},"7":{"start":{"line":9,"column":1},"end":{"line":28,"column":3}},"8":{"start":{"line":10,"column":2},"end":{"line":27,"column":3}},"9":{"start":{"line":11,"column":3},"end":{"line":11,"column":32}},"10":{"start":{"line":12,"column":18},"end":{"line":12,"column":46}},"11":{"start":{"line":13,"column":3},"end":{"line":15,"column":4}},"12":{"start":{"line":14,"column":4},"end":{"line":14,"column":29}},"13":{"start":{"line":16,"column":3},"end":{"line":21,"column":4}},"14":{"start":{"line":17,"column":4},"end":{"line":17,"column":36}},"15":{"start":{"line":20,"column":4},"end":{"line":20,"column":86}},"16":{"start":{"line":22,"column":3},"end":{"line":22,"column":101}},"17":{"start":{"line":23,"column":3},"end":{"line":23,"column":22}},"18":{"start":{"line":26,"column":3},"end":{"line":26,"column":58}},"19":{"start":{"line":32,"column":0},"end":{"line":32,"column":29}}},"fnMap":{"0":{"name":"(anonymous_0)","decl":{"start":{"line":3,"column":18},"end":{"line":3,"column":19}},"loc":{"start":{"line":3,"column":51},"end":{"line":30,"column":1}},"line":3},"1":{"name":"(anonymous_1)","decl":{"start":{"line":9,"column":14},"end":{"line":9,"column":15}},"loc":{"start":{"line":9,"column":24},"end":{"line":28,"column":2}},"line":9}},"branchMap":{"0":{"loc":{"start":{"line":10,"column":2},"end":{"line":27,"column":3}},"type":"if","locations":[{"start":{"line":10,"column":2},"end":{"line":27,"column":3}},{"start":{"line":10,"column":2},"end":{"line":27,"column":3}}],"line":10},"1":{"loc":{"start":{"line":13,"column":3},"end":{"line":15,"column":4}},"type":"if","locations":[{"start":{"line":13,"column":3},"end":{"line":15,"column":4}},{"start":{"line":13,"column":3},"end":{"line":15,"column":4}}],"line":13},"2":{"loc":{"start":{"line":16,"column":3},"end":{"line":21,"column":4}},"type":"if","locations":[{"start":{"line":16,"column":3},"end":{"line":21,"column":4}},{"start":{"line":16,"column":3},"end":{"line":21,"column":4}}],"line":16}},"s":{"0":1,"1":1,"2":1,"3":1,"4":1,"5":1,"6":1,"7":1,"8":1,"9":1,"10":1,"11":1,"12":0,"13":1,"14":1,"15":0,"16":1,"17":1,"18":0,"19":1},"f":{"0":1,"1":1},"b":{"0":[1,0],"1":[0,1],"2":[1,0]},"_coverageSchema":"43e27e138ebf9cfc5966b082cf9a028302ed4184","hash":"14f6f1e926bc831aea7248955e44b360943f663b","contentHash":"32851cef8213910b113f2e664cfe5b6b8984305e8ff9fcb9fa5af2b070a2c9ff"}}
