+++ .nyc_output/processinfo/012b8be2-e571-44a5-85a9-cef5da3aa102.json
@@ -1,0 +1,1 @@
+{"uuid":"012b8be2-e571-44a5-85a9-cef5da3aa102","parent":null,"pid":29342,"argv":["/usr/bin/node","/usr/local/bin/mocha","backend-test.js"],"execArgv":[],"cwd":"/usr/local/splash/test","time":1565922724635,"ppid":29335,"root":"d38cf1e3-cdbc-4859-8fab-eb6ff388e543","coverageFilename":"/usr/local/splash/.nyc_output/012b8be2-e571-44a5-85a9-cef5da3aa102.json","files":["/usr/local/splash/models/filedata.js","/usr/local/splash/models/makeDir.js","/usr/local/splash/models/renameFile.js","/usr/local/splash/models/deleteFile.js","/usr/local/splash/models/vodlist.js","/usr/local/splash/schemas/index.js","/usr/local/splash/schemas/vodmeta.js"]}
