--- .nyc_output/processinfo/0961476f-548a-483e-bdcb-44b14ef1203d.json
@@ -1,1 +1,0 @@
-{"uuid":"0961476f-548a-483e-bdcb-44b14ef1203d","parent":null,"pid":4345,"argv":["/usr/bin/node","/usr/local/bin/mocha","metadata.spec.js"],"execArgv":[],"cwd":"/usr/local/splash/test","time":1565676334543,"ppid":4339,"root":"e0db929f-5575-4396-9636-48a1712e28b3","coverageFilename":"/usr/local/splash/.nyc_output/0961476f-548a-483e-bdcb-44b14ef1203d.json","files":["/usr/local/splash/models/metaData.js"]}
