+++ .nyc_output/processinfo/index.json
@@ -1,1 +1,1 @@
-{"processes":{"0961476f-548a-483e-bdcb-44b14ef1203d":{"parent":null,"children":[]}},"files":{"/usr/local/splash/models/metaData.js":["0961476f-548a-483e-bdcb-44b14ef1203d"]},"externalIds":{}}
+{"processes":{"012b8be2-e571-44a5-85a9-cef5da3aa102":{"parent":null,"children":[]}},"files":{"/usr/local/splash/models/filedata.js":["012b8be2-e571-44a5-85a9-cef5da3aa102"],"/usr/local/splash/models/makeDir.js":["012b8be2-e571-44a5-85a9-cef5da3aa102"],"/usr/local/splash/models/renameFile.js":["012b8be2-e571-44a5-85a9-cef5da3aa102"],"/usr/local/splash/models/deleteFile.js":["012b8be2-e571-44a5-85a9-cef5da3aa102"],"/usr/local/splash/models/vodlist.js":["012b8be2-e571-44a5-85a9-cef5da3aa102"],"/usr/local/splash/schemas/index.js":["012b8be2-e571-44a5-85a9-cef5da3aa102"],"/usr/local/splash/schemas/vodmeta.js":["012b8be2-e571-44a5-85a9-cef5da3aa102"]},"externalIds":{}}
