+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "1.0.1": "^1.0.0",
     "bluebird": "^3.5.5",
+    "chai": "^4.2.0",
     "cors": "^2.8.5",
     "directory-tree": "^2.2.3",
     "express": "^4.17.1",
@@ -22,6 +23,7 @@
     "pug": "^2.0.4",
     "queue": "^6.0.1",
     "sequential-task-queue": "^1.2.1",
+    "sinon": "^7.4.1",
     "torrent": "^2.0.0",
     "version": "^0.1.2",
     "webtorrent": "^0.107.2"
