+++ routes/makeDir.js
@@ -1,5 +1,4 @@
 var makeDir = function(path, dirName){
-	console.log('makeDir');
 	var fs = require('fs');
 	fs.mkdirSync(path + '/' + dirName);
 }
@@ -8,7 +7,7 @@ var express = require('express');
 var router = express.Router();
 
 router.get('/', (req, res) => {
-	makeDir(req.param('path'), req.param('dirName'));
+	makeDir(req.params.path, req.params.dirName);
 	res.send();
 });
