+++ views/home.pug
@@ -46,6 +46,7 @@ html
     script(src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.2/jquery.min.js")
     script(src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/js/bootstrap.min.js")
     script(src="restFrontHome.js")
-    
- 
+    script(src="playVideo.js")
+    script(src="posterRender.js")
+    script(src="depthRender.js")
