+++ views/restFrontHome.js
@@ -5,18 +5,6 @@ var hostname = document.location.hostname;
 var library = ['Actors', 'Genre'];
 var depthStacks = new Array();
 
-renderDepth(library[0]);
-
-function gnbHTML(meta){
-	var div = document.createElement('div');
-	div.className = 'col-sm-3'
-	var button = document.createElement('button');
-	button.innerHTML = meta;
-	button.className = 'btn btn-primary btn-lg btn-block';
-	button.setAttribute('onclick', 'renderDepth(\'' + meta + '\')');
-	div.appendChild(button);
-	return div;
-}
 
 function goBack(){
 	if(depthStacks.length == 0){
@@ -32,90 +20,7 @@ function goBack(){
 	}
 }
 
-function renderDepth(filter){
-	var posters = document.getElementById('posters');
-	var gnb = document.getElementById('gnb');
-	if(!library.includes(filter) && library.length != depthStacks.length){
-		depthStacks.push(filter);
-	}
-	var json = new Object();
-	for(i in depthStacks){
-		json[library[i]] = depthStacks[i]
-	}
-	if(depthStacks.length == library.length){
-		gnb.innerHTML = 'VOD LIST';
-		renderPoster(json);
-		return;
-	}
-	gnb.innerHTML = library[depthStacks.length];
-	var xhr = new XMLHttpRequest();
-	xhr.open('GET', 'http://' + hostname + ':8390/getDepth?json=' + JSON.stringify(json) + '&distinct=' + library[depthStacks.length]);
-	xhr.withCredentials = true;
-	xhr.send();
-	xhr.onload = function() {
-		if (xhr.status = 200) {
-			var res = xhr.response.substr(1).slice(0, -1);
-			var gnb = res.split(',');
-			for(i in gnb){
-				gnb[i] = gnb[i].replace(/"/g, "");
-			}
-			var div = document.getElementById('p-row');
-			while(div.firstChild){
-				div.removeChild(div.firstChild);
-			}
-			for(i in gnb){
-				var button = gnbHTML(gnb[i]);
-				div.appendChild(button);
-			}
-		} 
-		else {
-			console.log('Error ${xhr.status} : ${xhr.statusText}');
-		}
-	};
-};
 
-function playVideo(){
-	var path = document.getElementById('Path').innerHTML;
-	console.log(path);
-	var xhr = new XMLHttpRequest();
-	xhr.open('GET', 'http://' + hostname + ':8390/requesthls' + '?vodpath=' + path);
-	xhr.withCredentials = true;
-	xhr.send();
-	xhr.onload = function() {
-		if (xhr.status = 200) {
-			if(document.getElementById('main-player') != null){
-				var fluidWrapper = document.getElementById('fluid_video_wrapper_main-player');
-				fluidWrapper.remove();
-			}
-			var hls = xhr.response;
-			var src = document.createElement('source');
-			src.src = hls;
-			src.type = 'application/x-mpegURL';
-			var video = document.createElement('video');
-			document.getElementById('modal-player-content').appendChild(video);
-			video.id = 'main-player';
-			video.appendChild(src);
-			fluidPlayer('main-player',
-				{
-					layoutControls:{
-						subtitlesEnabled: true,
-						layout: 'default',
-						playButtonShowing: false
-					}
-				}
-			);
-			var fluidWrapper = document.getElementById('fluid_video_wrapper_main-player');
-			fluidWrapper.style.width = '100%';
-			fluidWrapper.style.height= '100%';
-			$("#modal-synopsis").modal('hide');
-			$("#modal-player").modal('show');
-		} 
-		else {
-			console.log('Error ${xhr.status} : ${xhr.statusText}');
-		}
-	};
-
-}
 
 function appendPara(div, data){
 	var para = document.createElement('p');
@@ -159,51 +64,5 @@ function setModalMetaData(data){
 
 	//get request and change modal
 }
-
-function posterHTML(meta){
-	var div = document.createElement('div');
-	div.className = 'col-sm-3'
-	var img = document.createElement('img');
-	var str = new String(meta['Poster']);
-	if(str == 'N/A'){
-		img.src = 'http://103.117.231.226//Admin/main/no_poster.png'
-	}else{
-		img.src = meta['Poster'];
-	}
-	img.style.width = '300px'; img.style.height = '445px';
-	img.className = 'img-responsive';
-	img.dataset.target = '.bd-example-modal-lg';
-	img.dataset.toggle='modal';
-	img.setAttribute('onclick', 'setModalMetaData(\'' + meta['originalname'] + '\')');
-	div.appendChild(img);
-	appendPara(div, meta['Title']);
-	appendPara(div, meta['Year']);
-	appendPara(div, meta['originalname']);
-	return div;
-}
-
-function renderPoster(query='{}'){
-	var posters = document.getElementById('posters');
-	var xhr = new XMLHttpRequest();
-	xhr.open('GET', 'http://' + hostname + ':8390/synopsis?json=' + JSON.stringify(query));
-	xhr.withCredentials = true;
-	xhr.send();
-	xhr.onload = function() {
-		if (xhr.status = 200) {
-			var metadatas = JSON.parse(xhr.response);
-			var div = document.getElementById('p-row');
-			while(div.firstChild){
-				div.removeChild(div.firstChild);
-			}
-			for(i in metadatas){
-					var poster = posterHTML(metadatas[i]);
-					div.appendChild(poster);
-			}
-		} 
-		else {
-			console.log('Error ${xhr.status} : ${xhr.statusText}');
-		}
-	};
-};
